Fix issue with escaping in completion pattern
If the text you are trying to complete has a backslash (for something like a new
line) then it can lead to issues. Specifically a \ is represented by "\\" in
elisp strings. But \\ is invalid in a replacement text so you have to escape
that with \\\\. And because continued-match-fn was returning a string
replacement it had to be escaped again, leading to \\\\\\\\. Welcome to
backslash hell. It was easier to just change how to added the continued space
then mess with that.