Fix issue with csh on some linux distros
In the bindkey on some linux distros there are keybindings that don't play well
with Emacs and breaks the shell. These keybindings are only added if the
terminal is not "dumb", so it does not normally impact users. But if the
terminal is dumb, we can't use any keybindings to trigger completion (either \t
or ^D) so we need to set it to something else.

I am not going to change the name of the function (it mentions bash
specifically) because it will break backwards compatibility and I don't think
there are many users of the package now that `emacs-bash-completion` has gotten
so good. If you only use bash, that is the package you should be using instead
of this one.